move str_to_int to utility so it can be shared easily
